Move tracker to own sub-package
