+++ docs-i18n-tracker/.gitignore
@@ -1,3 +1,8 @@
+# build output
+dist/
+# generated types
+.astro/
+
 # dependencies
 node_modules/
 
@@ -6,6 +11,7 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 pnpm-debug.log*
+
 
 # environment variables
 .env
