+++ docs-i18n-tracker/scripts/lib/output.mjs
