+++ pnpm-workspace.yaml
@@ -2,3 +2,4 @@ packages:
   - 'packages/**/*'
   - 'examples/**/*'
   - 'docs'
+  - 'docs-i18n-tracker'
